wip gameMaster refactor
